README.md
The Lifting - Dev
The Lifting is a group game project for one of the courses during my studies in IT University of Copenhagen.
The game puts the player into the position of a window cleaner that primarily focuses on cleaning on the windows and discovering the dark truth behind the 
society's. The player will be able to track certain elements that will hint parts of the story. Each level has to be explored before moving to the next floor